refactor(playground): move provider media handling into its own package
Three changes, all aimed at the file upstream edits most.

Assert that media reaches each provider. The cross-provider test only checked that
each builder returned something, and with the media branch removed every builder
still returns a valid text-only message — so the check passed while the model
received no image. That is exactly the shape of a mis-resolved merge conflict in
one provider's branch. The assertions now read the payload: image_url for Chat
Completions, input_image for Responses, a base64 source for Anthropic, raw bytes
and a format token for Bedrock Converse.

Collapse the duplication. An identical 48-line block built OpenAI-style content in
three clients, and a further 7-line guard was repeated across the same three. Both
are now single functions. Copies are worse than their line count suggests: a change
applied to two of three leaves the third quietly sending something else.

Split media handling per provider under playground_media/. Each provider carries
media on its own wire format and those formats share nothing beyond needing bytes
and a media type, so separating them means a change to how Anthropic takes a
document cannot disturb Bedrock. It also keeps the code off upstream's path:
playground_clients.py is where new providers and models land, and the media
footprint there drops from 551 added lines to 86, with every remaining hunk a
one-to-three-line delegation instead of an interleaved block.
